Delay disposal of PyObject when creating a Tuple and moving from forced LINQ eval
Callined .ToArray on the IEnumerable will force it to evaluate in place, whereas if we don't do that, it will not evaluate the iterator until later on (or possibly never). But this means we can't dispose of the PyObject instances in the tupleValues collection, because we don't know when they'll be evaluated. Instead, leave it to the GC to deref them